185913628: Rename files to follow Rails conventions so they can be auto-loaded
