+++ app/controllers/concerns/pagination.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
-require_relative '../../services/pagination'
-
 module Pagination
   def self.paginate(collection:, params: {})
-    pagination = Services::Pagination.new(collection, params)
+    pagination = PaginationService.new(collection, params)
 
     [
       pagination.metadata,
