+++ app/view_models/pagination_view_model.rb
@@ -1,0 +1,48 @@
+# frozen_string_literal: true
+
+class PaginationViewModel
+  DEFAULT = { page: 1, per_page: 5 }.freeze
+  PER_PAGE = [5, 10, 25].freeze
+
+  attr_reader :page, :count, :per_page
+
+  def initialize(params = {})
+    @page = (params[:page] || DEFAULT[:page]).to_i
+    @count = params[:count]
+    @per_page = (params[:per_page] || DEFAULT[:per_page]).to_i
+  end
+
+  def offset
+    return 0 if page == 1
+
+    per_page * (page.to_i - 1)
+  end
+
+  def next_page
+    page + 1 unless last_page?
+  end
+
+  def next_page?
+    page < total_pages
+  end
+
+  def previous_page
+    page - 1 unless first_page?
+  end
+
+  def previous_page?
+    page > 1
+  end
+
+  def last_page?
+    page == total_pages
+  end
+
+  def first_page?
+    page == 1
+  end
+
+  def total_pages
+    (count / per_page.to_f).ceil
+  end
+end
